starting refactory of API
