+++ .gitignore
@@ -1,0 +1,5 @@
+Back_End/node_modules
+DB/
+
+
+
