+++ Back_End/api/models/authDAO.js
@@ -1,0 +1,7 @@
+function AuthDAO(connection) {
+    this._connection = connection ;   
+}
+
+AuthDAO.prototype.register = function (callback) {
+    this._connection.query("INSERT INTO users SET ?", userData,callback)
+}
