+++ Back_End/app.js
@@ -1,10 +1,8 @@
 var express = require('express');
 var cors = require('cors');
 var app = express();
-var auth = require('./Config/auth');
-var user =require('./Routes/users')
-var hello = require('./Routes/helloWorld')
-var port = process.env.PORT || 3000;
+var consign = require('consign');
+var port = process.env.PORT || 3050;
 
 //Loads the middlewares to parse these data format
 app.use(express.urlencoded({
@@ -15,11 +13,16 @@ app.use(cors());
 
 //Adding to use the users route
 
-app.use('/user', auth);
+// app.use('/user', auth);
 
-app.use('/user', user);
+// app.use('/user', user);
 
-app.use('/hello', hello);
+consign()
+    .include("./config")
+    .then("./api/routes/auth.js")
+    .then("./api/routes")
+    .then("./config")
+    .into(app)
 
 
 app.listen(port, function () {
