--- DB/Data/auto.cnf
@@ -1,2 +1,0 @@
-[auto]
-server-uuid=2a819adb-b890-11e8-9107-0242ac130002
